Moving variant_generator into suggest
