+++ python/ray/tune/__init__.py
@@ -7,7 +7,7 @@ from ray.tune.tune import run_experiments, Experiment
 from ray.tune.registry import register_env, register_trainable
 from ray.tune.result import TrainingResult
 from ray.tune.trainable import Trainable
-from ray.tune.variant_generator import grid_search
+from ray.tune.suggest import grid_search
 
 __all__ = [
     "Trainable", "TrainingResult", "TuneError", "grid_search", "register_env",
