+++ python/ray/tune/suggest/__init__.py
@@ -1,4 +1,11 @@
 from ray.tune.suggest.search import SearchAlgorithm, VariantAlgorithm
 from ray.tune.suggest.hyperopt import HyperOptSearch
+from ray.tune.suggest.variant_generator import generate_trials, grid_search
 
-__all__ = ["SearchAlgorithm", "VariantAlgorithm", "HyperOptSearch"]
+__all__ = [
+    "SearchAlgorithm",
+    "VariantAlgorithm",
+    "HyperOptSearch",
+    "grid_search",
+    "generate_trials"
+]
