+++ python/ray/tune/test/trial_runner_test.py
@@ -18,8 +18,8 @@ from ray.tune.experiment import Experiment
 from ray.tune.suggest.search import _MockAlgorithm, VariantAlgorithm
 from ray.tune.trial import Trial, Resources
 from ray.tune.trial_runner import TrialRunner
-from ray.tune.variant_generator import generate_trials, grid_search, \
-    RecursiveDependencyError
+from ray.tune.suggest import generate_trials, grid_search
+from ray.tune.suggest.variant_generator import RecursiveDependencyError
 
 
 class TrainableFunctionApiTest(unittest.TestCase):
